+++ tests/update_time/updaters/test_update_pyproject_toml.py
@@ -58,7 +58,9 @@ class UpdatePyprojectTomlsTest(LoggingTestCase):
         assert_success(update_pyproject_tomls())
         mock_pyproject_toml.write_text.assert_called_with('"package==1.1"\n')
         self.assert_path_logged(Path("uv.lock"))
-        self.assert_new_version_logged("package", "1.1, published: 2026-05-30 12:08")
+        self.assert_new_version_logged(
+            "package", "1.1, published: 2026-05-30 12:08", path=mock_pyproject_toml.relative_to()
+        )
         self.assert_no_warnings_logged()
 
     def test_update_with_changelog(self, run: Mock, get: Mock, glob: Mock):
@@ -73,7 +75,12 @@ class UpdatePyprojectTomlsTest(LoggingTestCase):
         assert_success(update_pyproject_tomls())
         mock_pyproject_toml.write_text.assert_called_with('"package_with_changelog==1.1"\n')
         self.assert_path_logged(Path("uv.lock"))
-        self.assert_new_version_logged("package_with_changelog", "1.1, published: 2026-05-30 12:07", self.changelog)
+        self.assert_new_version_logged(
+            "package_with_changelog",
+            "1.1, published: 2026-05-30 12:07",
+            self.changelog,
+            path=mock_pyproject_toml.relative_to(),
+        )
         self.assert_no_warnings_logged()
 
     def test_update_with_html_changelog(self, run: Mock, get: Mock, glob: Mock):
@@ -89,7 +96,9 @@ class UpdatePyprojectTomlsTest(LoggingTestCase):
         assert_success(update_pyproject_tomls())
         mock_pyproject_toml.write_text.assert_called_with('"package_with_html_changelog==1.1"\n')
         self.assert_path_logged(Path("uv.lock"))
-        self.assert_new_version_logged("package_with_html_changelog", "1.1, published: 2026-05-30 12:07")
+        self.assert_new_version_logged(
+            "package_with_html_changelog", "1.1, published: 2026-05-30 12:07", path=mock_pyproject_toml.relative_to()
+        )
         self.assert_no_warnings_logged()
 
     def test_update_with_github_url(self, run: Mock, get: Mock, glob: Mock):
@@ -106,7 +115,10 @@ class UpdatePyprojectTomlsTest(LoggingTestCase):
         mock_pyproject_toml.write_text.assert_called_with('"package_with_github_releases==1.1"\n')
         self.assert_path_logged(Path("uv.lock"))
         self.assert_new_version_logged(
-            "package_with_github_releases", "1.1, published: 2026-05-30 12:07", self.changelog
+            "package_with_github_releases",
+            "1.1, published: 2026-05-30 12:07",
+            self.changelog,
+            path=mock_pyproject_toml.relative_to(),
         )
         self.assert_no_warnings_logged()
 
@@ -119,7 +131,11 @@ class UpdatePyprojectTomlsTest(LoggingTestCase):
         assert_success(update_pyproject_tomls())
         mock_pyproject_toml.write_text.assert_called_with('"package_without_github_releases==1.1"\n')
         self.assert_path_logged(Path("uv.lock"))
-        self.assert_new_version_logged("package_without_github_releases", "1.1, published: 2026-05-30 12:07")
+        self.assert_new_version_logged(
+            "package_without_github_releases",
+            "1.1, published: 2026-05-30 12:07",
+            path=mock_pyproject_toml.relative_to(),
+        )
         self.assert_no_warnings_logged()
 
     def test_unchanged(self, run: Mock, get: Mock, glob: Mock):